converts activeTenant from string to object transparently
This fixes a bug that was introduced in 0.3.7, that release changed
activeTenant from a string to an object, which works perfectly for a
fresh install of the extension. However it broke for users who had set
their institution in a previous version of the extension and were
upgrading.

This change detects whether activeTenant stored in local storage is a
string and if so converts to an object after retrieving the tenant
details.